Clarify status message
If we examine a filename, and find a show name (or what we believe to be a show name) and query the provider for that show name, and get zero results, the UI reports, "Unable to download show information".  That makes it sound like the internet connection failed or something similar.  Change it to "Unable to find show information".